bwl: dwpal: dont stop on get driver param failure
Channel scan API  has dwell time input parameter .
Just before starting scan we have to change dwell time accordingly and
restore it after triggering the scan.
The problem is that it requires  get parameter directly from driver.

In case Its not supported or broken we cannot continue scanning.
Solution is to bypass setting the dwell time in case its not suppoted and
continue the channel scan.

Signed-off-by: Alex Kanter <alexx.kanter@intel.com>