mon_wlan_hal: set sequence on first scan results event
When a triggered scan completes, an empty 'Channel_Scan_Dump_Result' event
sent to signal the scan is completed and that results are ready.
At this point, if dump-results is requested, it will result with a
series of 'Channel_Scan_Dump_Result' events with the results.
Fix current logic by treating the first 'Channel_Scan_Dump_Result' as results-ready
and trigger the dump-result sequence.

Signed-off-by: Adam Dov <adamx.dov@intel.com>