mon_wlan_hal:fix handling of nlh->nlmsg_seq==0
the nlh->nlmsg_seq is 0 for "scan_results_ready" msg, this means
instead of failing the flow, we need to treat the msg as
"scan_results_ready" if m_nl_seq is 0.

after the change, for nlh->nlmsg_seq==0, if m_ln_seq==0 the msg is
treated as "scan_results_ready", otherwise as failure.

Signed-off-by: Adam Dov <adamx.dov@intel.com>