monitor_thread: fix a bug in dereferencing scan results event
Before the change, the sCHANNEL_SCAN_RESULTS_NOTIFICATION message was
dereferenced directly to channel_scan_results param, but then again
dereferenced. this, caused a crash when freeing the buffer at the end
of the scope.

After the change, the first dereference is done to the buffer and only
the second dereference is to the param inside the msg.

Signed-off-by: Adam Dov <adamx.dov@intel.com>