bwl: dwpal: fix wrong buffer initialization
Default initialization is wrong in this scenario.
results_buffer is actually a shared pointer to a uint8_t buffer.
So, calling a shared_ptr default constructor is actually writing zeros
to a buffer of an unknown size.

"results" on the other is a casted pointer to the results struct
that can be initialized using default constructor.

Signed-off-by: alex kanter <alexx.kanter@intel.com>