tlfv: Add dump results request and response
The dynamic channel scan uses a parameter called "Dwell time" to define
how much time the driver will "scan" any channel for incoming beacons.
Setting the dwell time to any value 0 or lower is considered invalid.
This is not true, since the dwell time can be set to 0.
Setting the dwell time to 0 will cause the scan trigger to be
pointless since it will spand no time in each channel listening for
beacons and the only channel whose neighboring aps will be cached is
the current active channel in the radio

To allow for a scan to run at dwell time 0 add the dump results
request and response messages in the controller and monitor.

The auto-generated code will be added to the follow-up commit.

Signed-off-by: Itay Elenzweig <itayx.elenzweig@intel.com>