dynamic_channel_selection_task: add scan dump request flow
The dynamic channel scan task manages the DCS feature, it is the
task's responsibility to trigger a scan when needed given the
parameters set in the DB.

When a scan is requested, check the Dwell time parameter,
normally the task should send a "trigger scan" request.
but on a dwell time = 0 scenario, it should sent a "scan dump"
request instead.

Add two new private functions to the DCS task
- dcs_request_scan_dump: used to launch the scan dump request
- dcs_request_scan_trigger: used to launch the scan trigger request
Change the flow of the TRIGGER_SCAN state case to use the two new
functions.

Signed-off-by: Itay Elenzweig <itayx.elenzweig@intel.com>