tlvf: split link metric query TLV into two TLVs
The Link Metric Query TLV contains an optional field. However, TLVF
does not currently support optional fields.

As a workaround we create two different TLVs, one that includes the
optional field and another one which does not. When parsing received
message, the application will check the length of included TLV to know
which one is it.

Signed-off-by: Mario Maz <mmazzapater@gmail.com>