common: bcl: remove version map
The version map is only useful if there are several modules with
different version. However, since we removed dynamic loading of
libraries using dlopen(), this no longer happens. Therefore, remove:
- BEEROCKS_INIT_SO_VERSION
- set_module_version()
- the module_name argument of get_module_version() etc.
- s_beerocks_version_map

The version obviously still has to be stored somewhere, so add
non-static members to the version class for the version attributes, and
add a static pointer to the version singleton object that is created by
BEEROCKS_INIT_BEEROCKS_VERSION. Fall back to the invalid strings if the
singleton object wasn't created (yet).

This change solves an issue with static linking. The
s_beerocks_version_map approach relied on the order of constructor calls
of static objects: the s_beerocks_version_map object must be initialised
before the s_beerocks_version object from BEEROCKS_INIT_BEEROCKS_VERSION
is constructed. However, the order of initialisation of static objects
is not defined. It just happens to work with dynamic libraries, because
the constructors of each library are called as soon as the library is
loaded. Since bcl is loaded before any library that constructs an
s_beerocks_version object is loaded, the order is OK. But with static
linking, this is no longer necessarily true.

With the singleton pointer, this problem no longer happens, because the
initialisation of POD objects is guaranteed to preceed user-defined
constructors.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>